Issue #325: fix startup-recovery test for cross-platform CI
Extract isProcessAlive into src/server/utils/process-utils.ts so tests
can mock it at the module boundary. The previous approach mocked
child_process.execSync, but the POSIX code path uses process.kill(pid, 0)
which was never intercepted — causing test failures on Linux CI.
